decimal amout

+++ Mat.Api/Dtos/IngredientDto.cs
@@ -6,6 +6,6 @@ public sealed record IngredientDto
 {
     public int Id { get; set; }
     public required string Name { get; set; }
-    public required int Amount { get; set; }
+    public required decimal Amount { get; set; }
     public Unit Unit { get; set; }
 }
